Removed item_window
