+++ glass-renderer/item.h
@@ -77,4 +77,8 @@ Item *item_get(int id);
 void item_add(Item *item);
 void item_remove(Item *item);
 
+extern void item_type_window_update_space_pos_from_window(Item *item);
+extern Item *item_get_from_window(Window window, int create);
+extern void items_get_from_toplevel_windows();
+
 #endif
