--- glass-renderer/item_window.c
@@ -1,115 +1,0 @@
-#include "glapi.h"
-#include "item_window.h"
-#include "item_window_shader.h"
-#include "item_window_pixmap.h"
-#include "item_window_svg.h"
-#include "property.h"
-#include "xapi.h"
-#include "wm.h"
-#include "debug.h"
-#include <X11/Xatom.h>
-
-void item_type_window_update_space_pos_from_window(Item *item) {
-  XWindowAttributes attr;
-  XGetWindowAttributes(display, item->window, &attr);
-
-  item->x                   = attr.x;
-  item->y                   = attr.y;
-  int width                 = attr.width;
-  int height                = attr.height;
-  DEBUG("window.spacepos", "Spacepos for %ld is %d,%d [%d,%d]\n", item->window, item->x, item->y, width, height);
-
-  item->width = width;
-  item->height = height;
-  
-  Atom type_return;
-  int format_return;
-  unsigned long nitems_return;
-  unsigned long bytes_after_return;
-  unsigned char *prop_return;
-  XGetWindowProperty(display, item->window, IG_COORDS, 0, sizeof(float)*4, 0, AnyPropertyType,
-                       &type_return, &format_return, &nitems_return, &bytes_after_return, &prop_return);
-  if (type_return != None) {
-    for (int i = 0; i < 4; i++) {
-     item->coords[i] = *(float *) (i + (long *) prop_return);
-    }
-    XFree(prop_return);
-  } else {
-    View *v = NULL;
-    if (views) {
-      v = view_find(views, item->layer);
-    }
-    if (v) {
-      item->coords[0] = v->screen[0] + (v->screen[2] * (float) item->x) / (float) v->width;
-      item->coords[1] = v->screen[1] + v->screen[3] - (v->screen[3] * (float) item->y) / (float) v->height;
-      item->coords[2] = (v->screen[2] * (float) item->width) / (float) v->width;
-      item->coords[3] = (v->screen[3] * (float) item->height) / (float) v->height;
-    } else {
-      item->coords[0] = ((float) (item->x - overlay_attr.x)) / (float) overlay_attr.width;
-      item->coords[1] = ((float) (overlay_attr.height - item->y - overlay_attr.y)) / (float) overlay_attr.width;
-      item->coords[2] = ((float) (width)) / (float) overlay_attr.width;
-      item->coords[3] = ((float) (height)) / (float) overlay_attr.width;
-    }
-    long arr[4];
-    for (int i = 0; i < 4; i++) {
-     arr[i] = *(long *) &item->coords[i];
-    }
-    XChangeProperty(display, item->window, IG_COORDS, XA_FLOAT, 32, PropModeReplace, (void *) arr, 4);
-  }
-}
-
-Item *item_get_from_window(Window window, int create) {
-  if (items_all) {
-    for (size_t idx = 0; idx < items_all->count; idx++) {
-      Item *item = (Item *) items_all->entries[idx];
-      if (item->window == window) return item;
-    }
-  }
-  if (!create) return NULL;
-  
-  DEBUG("window.add", "Adding window %ld\n", window);
-
-  Atom type_return;
-  int format_return;
-  unsigned long nitems_return;
-  unsigned long bytes_after_return;
-  unsigned char *prop_return;
-  
-  XGetWindowProperty(display, window, DISPLAYSVG, 0, 0, 0, XA_STRING, &type_return, &format_return, &nitems_return, &bytes_after_return, &prop_return);
-  XFree(prop_return);
-  if (type_return == None) {
-    return item_create(&item_type_window_pixmap, &window);
-  } else {
-    XGetWindowProperty(display, window, DISPLAYSVG, 0, bytes_after_return, 0, XA_STRING, &type_return, &format_return, &nitems_return, &bytes_after_return, &prop_return);
-    ItemWindowSVGArgs args = {window, (char *) prop_return};
-    return item_create(&item_type_window_svg, &args);
-  }
-}
-
-
-void items_get_from_toplevel_windows() {
-  XCompositeRedirectSubwindows(display, root, CompositeRedirectAutomatic);
-
-  XGrabServer(display);
-
-  Window returned_root, returned_parent;
-  Window* top_level_windows;
-  unsigned int num_top_level_windows;
-  XQueryTree(display,
-             root,
-             &returned_root,
-             &returned_parent,
-             &top_level_windows,
-             &num_top_level_windows);
-
-  for (unsigned int i = 0; i < num_top_level_windows; ++i) {
-    XWindowAttributes attr;
-    XGetWindowAttributes(display, top_level_windows[i], &attr);
-    if (attr.map_state == IsViewable) {
-      item_get_from_window(top_level_windows[i], True);
-    }
-  }
-
-  XFree(top_level_windows);
-  XUngrabServer(display);
-}
