+++ glass-renderer/xapi.c
@@ -1,5 +1,5 @@
 #include "xapi.h"
-#include "item_window.h"
+#include "item.h"
 #include <sys/types.h>
 #include <unistd.h>
 #include "error.h"
